feat: make multipayment transaction modals the default send action for the wallet
