+++ src/renderer/components/Transaction/index.js
@@ -1,9 +1,9 @@
 import TransactionAmount from './TransactionAmount'
 import TransactionDetail from './TransactionDetail'
 import TransactionModal from './TransactionModal'
-import TransactionMultiPaymentList from './TransactionMultiPaymentList'
 import TransactionPeerList from './TransactionPeerList'
 import TransactionStatusIcon from './TransactionStatusIcon'
+import TransactionTransferList from './TransactionTransferList'
 
 export * from './TransactionForm'
 
@@ -11,9 +11,9 @@ export {
   TransactionAmount,
   TransactionDetail,
   TransactionModal,
-  TransactionMultiPaymentList,
   TransactionPeerList,
-  TransactionStatusIcon
+  TransactionStatusIcon,
+  TransactionTransferList
 }
 
 export * from './TransactionShow'
