+++ src/renderer/components/Wallet/WalletHeading/WalletHeadingPrimaryActions.vue
@@ -62,55 +62,27 @@
       </template>
     </ButtonModal>
 
-    <ButtonDropdown
+    <ButtonModal
       v-show="!currentWallet.isContact"
-      :items="sendOptions"
-      classes="option-heading-button px-3 py-2"
-      class="mr-2"
+      :class="buttonStyle"
+      :label="$t('TRANSACTION.SEND')"
+      icon="send"
+      view-box="0 0 12 12"
     >
-      <ButtonModal
-        slot="primaryButton"
-        class="option-heading-button px-3 py-2 h-full"
-        :class="{
-          'rounded-tr-none': hasSendOptions,
-          'rounded-br-none': hasSendOptions
-        }"
-        :label="$t('TRANSACTION.SEND')"
-        icon="send"
-        view-box="0 0 12 12"
-      >
-        <template slot-scope="{ toggle, isOpen }">
-          <TransactionModal
-            v-if="isOpen"
-            :type="0"
-            @cancel="closeTransactionModal(toggle, isOpen)"
-            @sent="closeTransactionModal(toggle, isOpen)"
-          />
-        </template>
-      </ButtonModal>
-
-      <ButtonModal
-        slot="button"
-        slot-scope="{ item, triggerClose }"
-        :label="item.label"
-        class="option-heading-button whitespace-no-wrap w-full"
-        @toggle="triggerClose"
-      >
-        <template slot-scope="{ toggle, isOpen }">
-          <TransactionModal
-            v-if="isOpen"
-            :type="item.type"
-            @cancel="closeTransactionModal(toggle, isOpen)"
-            @sent="closeTransactionModal(toggle, isOpen)"
-          />
-        </template>
-      </ButtonModal>
-    </ButtonDropdown>
+      <template slot-scope="{ toggle, isOpen }">
+        <TransactionModal
+          v-if="isOpen"
+          :type="0"
+          @cancel="closeTransactionModal(toggle, isOpen)"
+          @sent="closeTransactionModal(toggle, isOpen)"
+        />
+      </template>
+    </ButtonModal>
   </div>
 </template>
 
 <script>
-import { ButtonDropdown, ButtonModal, ButtonReload } from '@/components/Button'
+import { ButtonModal, ButtonReload } from '@/components/Button'
 import { ModalQrCode } from '@/components/Modal'
 import { TransactionModal } from '@/components/Transaction'
 import { ContactRenameModal } from '@/components/Contact'
@@ -122,7 +94,6 @@ export default {
   inject: ['switchToTab', 'walletVote'],
 
   components: {
-    ButtonDropdown,
     ButtonModal,
     ButtonReload,
     ModalQrCode,
@@ -131,11 +102,9 @@ export default {
     SvgIcon
   },
 
-  data () {
-    return {
-      isRefreshing: false
-    }
-  },
+  data: () => ({
+    isRefreshing: false
+  }),
 
   computed: {
     buttonStyle () {
@@ -146,37 +115,8 @@ export default {
       return this.wallet_fromRoute
     },
 
-    currentNetwork () {
-      return this.session_network
-    },
-
     doesNotExist () {
       return !this.$store.getters['wallet/byAddress'](this.currentWallet.address)
-    },
-
-    hasAip11 () {
-      return this.currentNetwork.constants ? !!this.currentNetwork.constants.aip11 : false
-    },
-
-    sendOptions () {
-      const options = []
-
-      if (!this.hasAip11) {
-        return options
-      }
-
-      if (!this.currentWallet.isLedger) {
-        options.push({
-          label: this.$t('TRANSACTION.TYPE.MULTI_PAYMENT'),
-          type: 6
-        })
-      }
-
-      return options
-    },
-
-    hasSendOptions () {
-      return !!this.sendOptions.length
     },
 
     isVoting () {
